requirements.txt
pymongo[srv]
flask
passlib[bcrypt]
pyJWT[crypto]
python-dotenv